+++ Assets/Main/Analytics/Test Script/PlayerPosData3.cs
@@ -1,25 +1,21 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using UnityEngine;
 using System.IO;
-
+using UnityEngine;
 
 public class PlayerPosData3 : MonoBehaviour
 {
-    string content;
-    string path;
-    private List<Vector3> Positions = new List<Vector3>();
     public List<Vector3> savePositions = new List<Vector3>();
+    private readonly List<Vector3> Positions = new List<Vector3>();
     private Vector3 _currentPos = Vector3.zero;
+    private string content;
+    private string path;
 
     public void Awake()
     {
         path = Application.dataPath + "/LogPlayerPosData.txt";
-        if (!File.Exists(path))
-        {
-            File.WriteAllText(path, "");
-        }
+        if (!File.Exists(path)) File.WriteAllText(path, "");
 
         LoadPos();
     }
@@ -33,48 +29,40 @@ public class PlayerPosData3 : MonoBehaviour
         }
     }
 
-    void LoadPos()
+    public void Update()
     {
-        string in_data = File.ReadAllText(path);
-        Debug.Log($"in_data {in_data}");
-        string[] jsonString = in_data.Split('|');
-        foreach (var json in jsonString)
-        {
-            if (json.Length > 0)
-            {
-                Vector3 p = JsonUtility.FromJson<Vector3>(json);
-                Positions.Add(p);
-            }
-        }
+        savePositions = Positions;
     }
 
     private void OnDisable()
     {
         // testing saving data 
         Debug.Log($"Positions {Positions[0].ToString()}");
-        List<string> vectorJSON = new List<string>();
-        foreach (var vec in Positions)
-        {
-            vectorJSON.Add(JsonUtility.ToJson(vec));
-        }
+        var vectorJSON = new List<string>();
+        foreach (var vec in Positions) vectorJSON.Add(JsonUtility.ToJson(vec));
 
-        string _data = String.Join("|", vectorJSON);
+        var _data = string.Join("|", vectorJSON);
 
         //string dataAsJson = JsonUtility.ToJson(_data);
         File.AppendAllText(path, _data + '|');
     }
 
+    private void OnDestroy()
+    {
+        // cache the positions to disk
+    }
+
     private void OnDrawGizmos()
     {
-        Vector3 prev = Vector3.zero;
+        var prev = Vector3.zero;
         var c = Color.yellow;
         foreach (var position in Positions)
         {
             //this creates a box and line of the player's pos
             var p = position;
-            p.x = (int) p.x;
-            p.y = (int) p.y;
-            p.z = (int) p.z;
+            p.x = (int)p.x;
+            p.y = (int)p.y;
+            p.z = (int)p.z;
 
 
             c.a = 0.1f;
@@ -83,7 +71,7 @@ public class PlayerPosData3 : MonoBehaviour
 
 
             // testing saving
-            myClass myClass01 = new myClass();
+            var myClass01 = new myClass();
             myClass01.Positions2 = Positions;
 
             //allpositions = new Vector3(p.x, p.y, p.z);
@@ -93,34 +81,33 @@ public class PlayerPosData3 : MonoBehaviour
             Debug.Log(position);
         }
 
-        for (int i = 1; i < Positions.Count; i++)
+        for (var i = 1; i < Positions.Count; i++)
         {
             if (Vector3.Distance(Positions[i - 1], Positions[i]) > 2)
-            {
                 c = Color.black;
-            }
             else
-            {
                 c = Color.white;
-            }
 
             Debug.DrawLine(Positions[i - 1], Positions[i]);
         }
+    }
+
+    private void LoadPos()
+    {
+        var in_data = File.ReadAllText(path);
+        Debug.Log($"in_data {in_data}");
+        var jsonString = in_data.Split('|');
+        foreach (var json in jsonString)
+            if (json.Length > 0)
+            {
+                var p = JsonUtility.FromJson<Vector3>(json);
+                Positions.Add(p);
+            }
     }
 
     [Serializable]
     public class myClass
     {
         public List<Vector3> Positions2 = new List<Vector3>();
-    }
-
-    public void Update()
-    {
-        savePositions = Positions;
-    }
-
-    private void OnDestroy()
-    {
-        // cache the positions to disk
     }
 }
